typ-messenger-django/requirements.txt
asgi-redis==1.4.3
asgiref==1.1.2
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
certifi==2017.11.5
channels==1.1.8
chardet==3.0.4
constantly==15.1.0
daphne==1.3.0
defusedxml==0.5.0
Django==2.0
django-allauth==0.34.0
django-rest-auth==0.9.2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
msgpack-python==0.4.8
oauthlib==2.0.6
PyJWT==1.5.3
python3-openid==3.1.0
pytz==2017.3
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
urllib3==1.22
websocket-client==0.46.0
zope.interface==4.4.3
